v0.12.0

+++ app/src/main/java/com/cephmonitor/cephmonitor/layout/component/calculator/CountPercentageCalculator.java
@@ -97,7 +97,7 @@ public class CountPercentageCalculator extends OriginCalculator {
         this.minPercentage = minPercentage;
     }
 
-    public void setTotal(int total) {
+    public void setTotal(long total) {
         this.total = total;
         totalValue.setText(String.valueOf(total));
         part = (long) (total * maxPercentage);
